move files to logging file
